backend/requirements.txt
fastapi
uvicorn[standard]
torch
gymnasium[classic_control]
pillow
numpy
